Updated to 64bit mode. Multitasking is not yet updated.

+++ kernel/core/memory.h
@@ -5,7 +5,7 @@
 
 void mem_init();
 void mem_free(void* _addr);
-uint32_t mem_alloc(uint32_t size, char *name);
+uint64_t mem_alloc(uint32_t size, char *name);
 int mem_free_size();
 
 #endif
